+++ frontend/src/app/components/form-cars-create/form-cars-create.component.html
@@ -1,25 +1,43 @@
 <form [formGroup]="carForm" (ngSubmit)="onSubmit()">
   <div class="set-label-control">
     <label for="brand">Marca</label>
-    <select id="brand" formControlName="brand" (change)="getModels()">
+    <select id="brand" formControlName="brand" (change)="updateModels()">
       <option selected></option>
       @for (brand of brands(); track $index) {
         <option>{{ brand }}</option>
       }
     </select>
+    @if (
+      carForm.controls.brand.invalid &&
+      (carForm.controls.brand.dirty || carForm.controls.brand.touched)
+    ) {
+      @if (carForm.controls.brand.hasError("required")) {
+        <small>El campo es requerido</small>
+      }
+    }
   </div>
 
   <div class="set-label-control">
     <label for="model">Modelo</label>
     <select id="model" formControlName="model">
+      <option selected></option>
       @for (model of models(); track $index) {
         <option value="{{ model }}">{{ model }}</option>
       }
     </select>
+    @if (
+      carForm.controls.model.invalid &&
+      carForm.controls.model.value === "" &&
+      (carForm.controls.model.dirty || carForm.controls.model.touched)
+    ) {
+      @if (carForm.controls.model.hasError("required")) {
+        <small>El campo es requerido</small>
+      }
+    }
   </div>
 
   <h4>Detalles de los coches</h4>
-  <div class="cars-details" formArrayName="carDetails">
+  <div class="cars-details required" formArrayName="carDetails">
     @for (detail of carDetails.controls; track $index; let i = $index) {
       <div class="details-container" [formGroupName]="i">
         <h4>Detalle del coche {{ i + 1 }}</h4>
@@ -31,6 +49,25 @@
               id="registrationDate"
               formControlName="registrationDate"
             />
+            @if (
+              detail.controls.registrationDate.invalid &&
+              (detail.controls.registrationDate.dirty ||
+                detail.controls.registrationDate.touched)
+            ) {
+              @if (detail.controls.registrationDate.hasError("required")) {
+                <small>El campo es requerido</small>
+              }
+            }
+            @if (
+              detail.hasError("registrationDateIncorrect") &&
+              (detail.controls.registrationDate.dirty ||
+                detail.controls.registrationDate.touched)
+            ) {
+              <small>
+                La fecha de registro no puede ser anterior a la fecha de
+                fabricación</small
+              >
+            }
           </div>
 
           <div class="set-label-control">
@@ -40,6 +77,34 @@
               id="manufactureYear"
               formControlName="manufactureYear"
             />
+            @if (
+              detail.controls.manufactureYear.invalid &&
+              (detail.controls.manufactureYear.dirty ||
+                detail.controls.manufactureYear.touched)
+            ) {
+              @if (detail.controls.manufactureYear.hasError("required")) {
+                <small>El campo es requerido</small>
+              } @else if (detail.controls.manufactureYear.hasError("min")) {
+                <small
+                  >El año no puede ser menor que
+                  {{
+                    detail.controls.manufactureYear.errors!["min"]["min"]
+                  }}</small
+                >
+              } @else if (detail.controls.manufactureYear.hasError("max")) {
+                <small>El año no puede ser mayor que el año actual</small>
+              }
+            }
+            @if (
+              detail.hasError("registrationDateIncorrect") &&
+              (detail.controls.manufactureYear.dirty ||
+                detail.controls.manufactureYear.touched)
+            ) {
+              <small>
+                La fecha de registro no puede ser anterior a la fecha de
+                fabricación</small
+              >
+            }
           </div>
         </div>
 
@@ -47,6 +112,19 @@
           <div class="set-label-control">
             <label for="price">Precio</label>
             <input type="number" id="price" formControlName="price" />
+            @if (
+              detail.controls.price.invalid &&
+              (detail.controls.price.dirty || detail.controls.price.touched)
+            ) {
+              @if (detail.controls.price.hasError("required")) {
+                <small>El campo es requerido</small>
+              } @else if (detail.controls.price.hasError("min")) {
+                <small
+                  >Introducir un valor como mínimo de
+                  {{ detail.controls.price.errors!["min"]["min"] }}
+                </small>
+              }
+            }
           </div>
 
           <div class="set-label-control">
@@ -57,6 +135,15 @@
                 <option>{{ currency }}</option>
               }
             </select>
+            @if (
+              detail.controls.currency.invalid &&
+              (detail.controls.currency.dirty ||
+                detail.controls.currency.touched)
+            ) {
+              @if (detail.controls.currency.hasError("required")) {
+                <small>El campo es requerido</small>
+              }
+            }
           </div>
         </div>
 
@@ -68,17 +155,37 @@
               id="licensePlate"
               formControlName="licensePlate"
             />
+            @if (
+              detail.controls.licensePlate.invalid &&
+              (detail.controls.licensePlate.dirty ||
+                detail.controls.licensePlate.touched)
+            ) {
+              @if (detail.controls.licensePlate.hasError("required")) {
+                <small>El campo es requerido</small>
+              } @else if (detail.controls.licensePlate.hasError("pattern")) {
+                <small>Formato de matrícula incorrecto</small>
+              }
+            }
           </div>
 
           <div class="set-label-control">
             <label for="mileage">Kilometraje</label>
             <input type="number" id="mileage" formControlName="mileage" />
+            @if (
+              detail.controls.mileage.invalid &&
+              (detail.controls.mileage.dirty || detail.controls.mileage.touched)
+            ) {
+              @if (detail.controls.mileage.hasError("required")) {
+                <small>El campo es requerido</small>
+              }
+            }
           </div>
         </div>
 
         <div class="row-details">
           <div class="set-label-control">
             <label for="">Disponibilidad</label>
+
             <input
               type="checkbox"
               id="availability"
